refactor: isolate continuation trajectory artifacts
